Upgrade Clang Tidy

+++ src/ResultWriter/FaultWriter.cpp
@@ -71,10 +71,10 @@ void seissol::writer::FaultWriter::init(const unsigned int* cells,
   const AsyncCellIDs<3> cellIds(nCells, nVertices, cells, seissolInstance);
 
   // Create mesh buffers
-  bufferId = addSyncBuffer(cellIds.cells(), nCells * 3 * sizeof(int));
+  bufferId = addSyncBuffer(cellIds.cells(), static_cast<unsigned long>(nCells * 3) * sizeof(int));
   assert(bufferId == FaultWriterExecutor::Cells);
   NDBG_UNUSED(bufferId);
-  bufferId = addSyncBuffer(vertices, nVertices * 3 * sizeof(double));
+  bufferId = addSyncBuffer(vertices, static_cast<unsigned long>(nVertices * 3) * sizeof(double));
   assert(bufferId == FaultWriterExecutor::Vertices);
   NDBG_UNUSED(bufferId);
